Move entry transformers to menu package
