+++ src/main/kotlin/io/github/freya022/botcommands/api/pagination/menu/transformer/StringTransformer.kt
@@ -1,4 +1,4 @@
-package io.github.freya022.botcommands.api.pagination.transformer
+package io.github.freya022.botcommands.api.pagination.menu.transformer
 
 class StringTransformer : EntryTransformer<Any> {
     override fun toString(entry: Any): String = entry.toString()
